MXS-3107 Adapt to Columnstore cluster changes
Now at each monitor loop, the cluster configuration is also checked
and the internal server view is adjusted accordingly.

For full functionality, Columnstore must update the revision number
of the configuration at each change (something it currently does
not do).

It's somewhat messy to combine both the cluster sniffing an the
setting of the status bits, so those operations will be separated
in a subsequent commit. That will basically lead to some extra
work, but since a cluster change is not that common anyway, the
cluster check can and should be done less frequently than the
status check.