MXS-3178: Add WebSocket upgrade code
The REST API can now initiate the upgrade required to open WebSocket
connections. Once the upgrade is complete, the application (i.e. MaxScale)
takes over the communication.

The actual framing of the messages needs to be implemented separately as
the upgrade mechanism in libmicrohttpd is generic and not specific to
WebSockets.